Work towards a conduit interface.
This is really perverse code. I'm not sure the whole conduit idea is a good one, both
in general and in this specific case - although it would be idea for a log-munging daemon,
hS3 is conduit-free, and it looks like I'll have fiddle with a few types to get things to
compose. Oh the irony: conduits promise ease of composition, but actually break compositionality.

Known flaws with the code I've written:

- liftToEither totally breaks order - all Left values are passed straight through, while others
are treated normally.

- attoparsec-qua-conduit function requires a very well-behaved and clever parser.